REF: share _maybe_mask_result
