+++ pandas/core/arrays/boolean.py
@@ -23,7 +23,6 @@ from pandas._typing import (
 
 from pandas.core.dtypes.common import (
     is_bool_dtype,
-    is_float,
     is_float_dtype,
     is_integer_dtype,
     is_list_like,
@@ -531,36 +530,6 @@ class BooleanArray(BaseMaskedArray):
             )
 
         return self._maybe_mask_result(result, mask, other, op_name)
-
-    def _maybe_mask_result(self, result, mask, other, op_name: str):
-        """
-        Parameters
-        ----------
-        result : array-like
-        mask : array-like bool
-        other : scalar or array-like
-        op_name : str
-        """
-        # if we have a float operand we are by-definition
-        # a float result
-        # or our op is a divide
-        if (is_float_dtype(other) or is_float(other)) or (
-            op_name in ["rtruediv", "truediv"]
-        ):
-            from pandas.core.arrays import FloatingArray
-
-            return FloatingArray(result, mask, copy=False)
-
-        elif is_bool_dtype(result):
-            return BooleanArray(result, mask, copy=False)
-
-        elif is_integer_dtype(result):
-            from pandas.core.arrays import IntegerArray
-
-            return IntegerArray(result, mask, copy=False)
-        else:
-            result[mask] = np.nan
-            return result
 
     def __abs__(self):
         return self.copy()
